Removing more unsupported code from examples
pet.cpp has some file IO we should eventually support.

+++ data/examples/if1.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main()
 {
-   double d = 0.0;
+   int d = 0;
    d++;
    cout << d << endl;
 }
